+++ src/lib.rs
@@ -7,11 +7,11 @@
 //! development, especially focused on the error handling.
 //!
 //! # Motivation
-//! Error handling in proc-macros is unideal, as the top level functions of proc
-//! macros can only return `TokenStreams` both in success and failure case. This
-//! means that I often write code like this, moving the actual implementation in
-//! a separate function to be able to use the ergonomic rust error handling with
-//! e.g., `?`.
+//! Error handling in proc-macros is unideal, as the top level functions of
+//! proc-macros can only return `TokenStreams` both in success and failure case.
+//! This means that I often write code like this, moving the actual
+//! implementation in a separate function to be able to use the ergonomic rust
+//! error handling with e.g., `?`.
 //! ```
 //! # use proc_macro2::TokenStream;
 //! # use quote::quote;
@@ -37,7 +37,7 @@
 //!
 //! # Using the `#[manyhow]` macro
 //! To activate the error hadling, just add [`#[manyhow]`](manyhow) above any
-//! proc macro implementation, reducing the above example to:
+//! proc-macro implementation, reducing the above example to:
 //!
 //! ```
 //! # use quote::quote;
@@ -53,6 +53,17 @@
 //! fn my_macro(input: TokenStream2) -> syn::Result<TokenStream2> {
 //!     // ..
 //! #   Ok(quote!())
+//! }
+//!
+//! // On top of the TokenStreams any type that implements `syn::Parse` is supported
+//! # let _ = quote!{
+//! #[manyhow(proc_macro_derive(MyMacro))]
+//! #[proc_macro]
+//! # };
+//! // The output can also be anything that implements `quote::ToTokens`
+//! fn my_derive_macro(input: syn::DeriveInput) -> manyhow::Result<syn::ItemImpl> {
+//!     // ..
+//! #   manyhow::bail!("hello")
 //! }
 //! ```
 //!
@@ -112,7 +123,7 @@
 //! pub fn actual_macro(input: TokenStream2) -> TokenStream2 {
 //!     // ...
 //! }
-//! // would expand to
+//! // would roughly expand to
 //! #[proc_macro]
 //! pub fn actual_macro(input: TokenStream) -> TokenStream {
 //!     actual_macro_impl(input.into()).into()
@@ -127,8 +138,12 @@
 //! adding `manyhow` with `default-features=false`.
 //!
 //! The usage is more or less the same, though with some added boilerplate from
-//! needing to invoke one of [`function`], [`attribute`] or [`derive`](derive())
-//! directly.
+//! needing to invoke one of [`function()`] ([`function!`]), [`attribute()`]
+//! ([`attribute!`]) or [`derive()`] ([`derive!`]) directly. For each version
+//! there exists a function and a `macro_rules` macro, while the function only
+//! supports [`proc_macro::TokenStream`] and [`proc_macro2::TokenStream`], the
+//! macro versions also support any type that implements [`Parse`]
+//! and [`ToTokens`] respectively.
 //!
 //! While the examples use closures, functions can be passed in as well. The
 //! above example would then change to:
@@ -142,6 +157,8 @@
 //! #[proc_macro]
 //! # };
 //! pub fn my_macro(input: TokenStream) -> TokenStream {
+//! # let tmp = input.clone();
+//! # let output: TokenStream = 
 //!     manyhow::function(
 //!         input,
 //!         false,
@@ -150,12 +167,23 @@
 //! #           Ok(quote!())
 //!         },
 //!     )
+//! # ;
+//! # let input = tmp;
+//!     // Or
+//!     manyhow::function!(
+//!         input,
+//!         |input: syn::DeriveInput| -> manyhow::Result<syn::ItemImpl> {
+//!             // ..
+//! #           manyhow::bail!("error")
+//!         },
+//!     )
 //! }
 //! ```
 //! [`Emitter`](#emitter-mut-emitter) and [dummy
-//! `TokenStream`](#dummy-mut-tokenstream) can also be used. [`function`] and
-//! [`attribute`] take an additional boolean parameter controlling whether the
-//! input/item will be used as initial dummy.
+//! `TokenStream`](#dummy-mut-tokenstream) can also be used. [`function()`]
+//! ([`function!`]) and [`attribute()`] ([`attribute!`]) take an additional
+//! boolean parameter controlling whether the input/item will be used as initial
+//! dummy.
 //!
 //! # `emitter: &mut Emitter`
 //! [`MacroHandler`]s (the trait defining what closures/functions can be used
@@ -199,10 +227,11 @@
 
 use std::convert::Infallible;
 
+#[cfg(feature = "macros")]
 pub use macros::manyhow;
 use proc_macro2::TokenStream;
 #[cfg(doc)]
-use quote::ToTokens;
+use {quote::ToTokens, syn2::parse::Parse};
 
 extern crate proc_macro;
 
@@ -214,17 +243,98 @@ mod macro_rules;
 mod error;
 pub use error::*;
 
+mod parse_to_tokens;
+
 #[doc(hidden)]
 pub mod __private {
+    pub use std::prelude::rust_2021::*;
+
+    use proc_macro2::TokenStream;
+    pub use quote;
+
     pub use crate::span_ranged::{SpanRangedToSpanRange, ToTokensToSpanRange};
+    pub type Dummy = Option<TokenStream>;
+
+    pub use crate::parse_to_tokens::*;
 }
 
 /// Marker trait for [`proc_macro::TokenStream`] and
 /// [`proc_macro2::TokenStream`]
-pub trait AnyTokenStream: Clone + From<TokenStream> + Into<TokenStream> {}
+pub trait AnyTokenStream: Clone + From<TokenStream> + Into<TokenStream> + Default {}
 impl AnyTokenStream for TokenStream {}
 impl AnyTokenStream for proc_macro::TokenStream {}
 
+macro_rules! handler {
+    ($(#$doc:tt)*$name:ident; $($input:ident: $Input:ident),*; $($dummy:ident, $dummy_value:ident)?) => {
+        $(#$doc)*
+        pub fn $name<
+            $($Input: AnyTokenStream,)*
+            Output: AnyTokenStream,
+            Return: AnyTokenStream,
+            Error: ToTokensError,
+            Function,
+        >(
+            $($input: impl AnyTokenStream,)*
+            $($dummy: bool,)?
+            body: impl MacroHandler<($($Input,)*), Output, Output, Function, Error>,
+        ) -> Return {
+            #[allow(unused_mut)]
+            let mut tokens = Output::default();
+            $(let mut tokens = if $dummy {
+                $dummy_value.clone().into().into()
+            } else {
+                tokens
+            };)?
+            let mut emitter = Emitter::new();
+            let output = body.call(($($input.into().into(),)*), &mut tokens, &mut emitter);
+            let mut tokens = tokens.into();
+            let mut tokens = match output {
+                Ok(tokens) => tokens.into(),
+                Err(error) => {
+                    error.to_tokens(&mut tokens);
+                    tokens
+                }
+            };
+            emitter.to_tokens(&mut tokens);
+            tokens.into()
+        }
+    };
+}
+
+#[macro_export]
+#[doc(hidden)]
+macro_rules! __macro_handler {
+    ($name:ident; $($(#attr=$attr:tt)? $n:ident: $input:expr),+; $impl:expr$(; dummy:$dummy:expr)?) => {
+        $crate::__macro_handler! {! $name; $($(#attr=$attr)? $n: $input.clone()),+; $impl $(; $crate::__private::Some($dummy))?}
+    };
+    ($name:ident; $($(#attr=$attr:tt)? $n:ident: $input:expr),+; $impl:expr; dummy) => {
+        $crate::__macro_handler! {! $name; $($(#attr=$attr)? $n: $input),+; $impl; $crate::__private::Dummy::None}
+    };
+    (! $name:ident; $($(#attr=$attr:tt)? $n:ident: $input:expr),+; $impl:expr $(; $dummy:expr)?) => {{
+        use $crate::__private::ParseToTokens;
+        let implementation = $impl;
+        $(let $n = &$crate::__private::WhatType::new();)+
+        if false {
+            _ = $crate::__private::$name($($n.identify(),)+ $($dummy,)? implementation);
+            unreachable!();
+        } else {
+            match $crate::__private::$name($(
+                {#[allow(unused)]
+                let attr = false;
+                $(let attr = $attr;)?
+                $n.manyhow_parse($input, attr)},
+            )+ $($dummy,)? implementation)
+            {
+                Err(tokens) => tokens.into(),
+                Ok((output, tokens)) => (&$crate::__private::WhatType::from(&output))
+                    .manyhow_into_token_stream(output, tokens)
+                    .into(),
+            }
+        }
+    }};
+}
+
+handler! {
 /// Handles [`proc_macro_attribute`](https://doc.rust-lang.org/reference/procedural-macros.html#attribute-macros)
 /// implementation
 ///
@@ -289,44 +399,85 @@ impl AnyTokenStream for proc_macro::TokenStream {}
 ///
 /// assert_tokens! {output, {struct Struct(HelloWorld);}};
 /// ```
-pub fn attribute<
-    Input: AnyTokenStream,
-    Item: AnyTokenStream,
-    Output: AnyTokenStream,
-    Return: AnyTokenStream,
-    Error: ToTokensError,
-    Function,
->(
-    input: impl AnyTokenStream,
-    item: impl AnyTokenStream,
-    item_as_dummy: bool,
-    body: impl MacroHandler<(Input, Item), Output, Function, Error>,
-) -> Return {
-    let mut tokens: Output = if item_as_dummy {
-        item.clone().into().into()
-    } else {
-        TokenStream::default().into()
+attribute; input: Input, item: Item; item_as_dummy, item
+}
+
+/// Handles [`proc_macro_attribute`](https://doc.rust-lang.org/reference/procedural-macros.html#attribute-macros)
+/// implementation
+///
+/// Takes any `TokenStream` for `input` and `item` and its return value. If
+/// `#[as_dummy]` is specified on item, it will be used as default
+/// dummy code on error. `body` takes a [`MacroHandler`] with two `TokenStream`
+/// or type implementing [`Parse`] parameters and returning a `TokenStream` or
+/// type implementing [`ToTokens`]. And an optional [`&mut Emitter`](Emitter)
+/// and a `&mut TokenStream` for storing a dummy output.
+///
+///
+/// ```
+/// # use proc_macro_utils::assert_tokens;
+/// # use quote::{quote, ToTokens};
+/// use manyhow::{attribute, Emitter, Result};
+/// use proc_macro2::TokenStream;
+/// # let input = quote!();
+/// # let item = quote!();
+/// # let output: TokenStream =
+/// attribute!(input, item, |input: TokenStream,
+///                          item: TokenStream,
+///                          dummy: &mut TokenStream,
+///                          emitter: &mut Emitter|
+///  -> Result {
+///     // ..
+///         # Ok(quote!())
+/// });
+/// ```
+///
+/// *Note:* When `#[as_dummy]` is specified the `dummy: &mut TokenStream` will
+/// be initialized with `item`. To override assign a new `TokenStream`:
+/// ```
+/// # use proc_macro_utils::assert_tokens;
+/// # use syn2 as syn;
+/// use manyhow::{attribute, Result, SilentError};
+/// use proc_macro2::TokenStream;
+/// use quote::{quote, ToTokens};
+/// # let input = quote!(input);
+/// let item = quote!(
+///     struct Struct;
+/// );
+/// let output: TokenStream = attribute!(
+///     input,
+///     #[as_dummy]
+///     item,
+///     |input: TokenStream,
+///      item: syn::ItemStruct,
+///      dummy: &mut TokenStream|
+///      -> Result<syn::ItemStruct, SilentError> {
+///         assert_tokens!(dummy.to_token_stream(), {
+///             struct Struct;
+///         });
+///         *dummy = quote! {
+///             struct Struct(HelloWorld);
+///         };
+///         // ..
+///         Err(SilentError)
+///     },
+/// );
+///
+/// assert_tokens! {output, {struct Struct(HelloWorld);}};
+/// ```
+#[macro_export]
+macro_rules! attribute {
+    ($input:expr, #[as_dummy] $item:expr, $impl:expr $(,)?) => {
+        $crate::__macro_handler!{attribute_transparent; #attr=true input: $input, item: $item.clone(); $impl; dummy: $item}
     };
-    let mut emitter = Emitter::new();
-    let output = body.call(
-        (input.into().into(), item.into().into()),
-        &mut tokens,
-        &mut emitter,
-    );
-    let mut tokens = tokens.into();
-    let mut tokens = match output {
-        Ok(tokens) => tokens.into(),
-        Err(error) => {
-            error.to_tokens(&mut tokens);
-            tokens
-        }
+    ($input:expr, $item:expr, $impl:expr $(,)?) => {
+        $crate::__macro_handler!{attribute_transparent; #attr=true input: $input, item: $item; $impl; dummy}
     };
-    emitter.to_tokens(&mut tokens);
-    tokens.into()
 }
 
 /// Handles [`proc_macro_derive`](https://doc.rust-lang.org/reference/procedural-macros.html#derive-macros)
-/// implementation
+/// implementation.
+///
+/// Use [`derive!`] to support [`Parse`] and [`ToTokens`] as well.
 ///
 /// Takes any `TokenStream` for `item` and returns any `TokenStream`. `body`
 /// takes a [`MacroHandler`] with one `TokenStream` parameter. And an optional
@@ -356,7 +507,7 @@ pub fn derive<
     Function,
 >(
     item: impl AnyTokenStream,
-    body: impl MacroHandler<(Item,), Output, Function, Error>,
+    body: impl MacroHandler<(Item,), Output, Output, Function, Error>,
 ) -> Return {
     let mut tokens = TokenStream::default().into();
     let mut emitter = Emitter::new();
@@ -373,8 +524,42 @@ pub fn derive<
     tokens.into()
 }
 
+/// Handles [`proc_macro_derive`](https://doc.rust-lang.org/reference/procedural-macros.html#derive-macros)
+/// implementation.
+///
+/// Takes any `TokenStream` for `item` and returns any `TokenStream`. `body`
+/// takes a [`MacroHandler`] with one `TokenStream` or type implementing
+/// [`Parse`] parameter and returns a `TokenStream` or type implementing
+/// [`ToTokens`]. And an optional [`&mut Emitter`](Emitter) and `&mut
+/// TokenStream` for storing a dummy output.
+///
+/// ```
+/// # use proc_macro_utils::assert_tokens;
+/// # use quote::{quote, ToTokens};
+/// # use syn2 as syn;
+/// use manyhow::{derive, Emitter, Result};
+/// use proc_macro2::TokenStream;
+/// # let item = quote!();
+/// # let output: TokenStream =
+/// derive!(item, |item: syn::DeriveInput,
+///                dummy: &mut TokenStream,
+///                emitter: &mut Emitter|
+///  -> Result {
+///     // ..
+///         # Ok(quote!())
+/// });
+/// ```
+#[macro_export]
+macro_rules! derive {
+    ($item:expr, $impl:expr $(,)?) => {
+        $crate::__macro_handler! {derive_transparent; item: $item; $impl}
+    };
+}
+
 /// Handles function like [`proc_macro`](https://doc.rust-lang.org/reference/procedural-macros.html#function-like-procedural-macros)
 /// implementation
+///
+/// Use [`function!`] to support [`Parse`] and [`ToTokens`] as well.
 ///
 /// Takes any `TokenStream` for `input` and returns any
 /// `TokenStream`. If `input_as_dummy = true` the item input will be used as
@@ -435,7 +620,7 @@ pub fn function<
 >(
     input: impl AnyTokenStream,
     input_as_dummy: bool,
-    body: impl MacroHandler<(Input,), Output, Function, Error>,
+    body: impl MacroHandler<(Input,), Output, Output, Function, Error>,
 ) -> Return {
     let mut tokens: Output = if input_as_dummy {
         input.clone().into().into()
@@ -456,6 +641,104 @@ pub fn function<
     tokens.into()
 }
 
+/// Handles function like [`proc_macro`](https://doc.rust-lang.org/reference/procedural-macros.html#function-like-procedural-macros)
+/// implementation
+///
+/// Takes any `TokenStream` for `input` and returns any `TokenStream`. If
+/// `#[as_dummy]` is specified on input, it will be used as default
+/// dummy code on error. `body` takes a [`MacroHandler`] with one `TokenStream`
+/// or type implementing [`Parse`] parameter and returns a `TokenStream` or type
+/// implementing [`ToTokens`]. And an optional [`&mut Emitter`](Emitter) and a
+/// `&mut TokenStream` for storing a dummy output.
+///
+/// ```
+/// # use proc_macro_utils::assert_tokens;
+/// # use quote::{quote, ToTokens};
+/// # use syn2 as syn;
+/// use manyhow::{function, Emitter, Result};
+/// use proc_macro2::TokenStream;
+/// # let input = quote!();
+/// # let output: TokenStream =
+/// function!(input, |input: syn::Item,
+///                   dummy: &mut TokenStream,
+///                   emitter: &mut Emitter|
+///  -> Result<syn::ItemImpl> {
+///     // ..
+///         # manyhow::bail!("unimplemented")
+/// });
+/// ```
+///
+/// *Note:* When `#[as_dummy]` is specified on the input, the `dummy: &mut
+/// TokenStream` will be initialized with `input`. To override assign a new
+/// `TokenStream`:
+///
+/// ```
+/// use proc_macro_utils::assert_tokens;
+/// use manyhow::{function, Result, SilentError};
+/// use proc_macro2::TokenStream;
+/// use quote::{quote, ToTokens};
+///
+/// let input = quote!(some input);
+/// let output: TokenStream = function!(
+///     #[as_dummy] input,
+///     |input: TokenStream, dummy: &mut TokenStream|
+///      -> Result<TokenStream, SilentError> {
+///         assert_tokens!(dummy.to_token_stream(), {
+///             some input
+///         });
+///         *dummy = quote! {
+///             another input
+///         };
+///         // ..
+///         Err(SilentError)
+///     },
+/// );
+///
+/// assert_tokens! {output, {another input}};
+/// ```
+#[macro_export]
+macro_rules! function {
+    (#[as_dummy] $input:expr, $impl:expr $(,)?) => {
+        $crate::__macro_handler! {function_transparent; input: $input; $impl; dummy: $input}
+    };
+    ($input:expr, $impl:expr $(,)?) => {
+        $crate::__macro_handler! {function_transparent; input: $input; $impl; dummy}
+    };
+}
+
+#[test]
+fn function_macro() {
+    use proc_macro::TokenStream as TokenStream1;
+    use quote::quote;
+    // proc_macro2::TokenStream
+    let output: TokenStream =
+        function!(quote!(hello), |input: TokenStream| -> TokenStream { input });
+    assert_eq!(output.to_string(), "hello");
+    // proc_macro::TokenStream do not run :D
+    if false {
+        let _: TokenStream1 = function!(
+            TokenStream1::from(quote!(hello)),
+            |input: TokenStream1| -> TokenStream1 { input }
+        );
+    }
+
+    #[cfg(feature = "syn2")]
+    {
+        use quote::ToTokens;
+        let output: TokenStream = function!(#[as_dummy] quote!(hello;), |input: syn2::LitInt| -> TokenStream {
+            input.into_token_stream()
+        });
+        assert_eq!(
+            output.to_string(),
+            quote!(hello; ::core::compile_error! { "expected integer literal" }).to_string()
+        );
+        let output: TokenStream = function!(quote!(20), |_input: syn2::LitInt| -> syn2::Ident {
+            syn2::parse_quote!(hello)
+        });
+        assert_eq!(output.to_string(), "hello");
+    }
+}
+
 /// Implementation of a proc-macro
 ///
 /// Note: for `TokenStream` either [`proc_macro::TokenStream`] or
@@ -470,31 +753,32 @@ pub fn function<
 /// Emitter`](Emitter) which allows emitting errors without returning early. And
 /// a `&mut TokenStream` to return a dummy `TokenStream` on failure, note that
 /// this `TokenStream` must be the same type as the one returned.
-pub trait MacroHandler<Input, Output, Function, Error = Infallible> {
+pub trait MacroHandler<Input, Dummy, Output, Function, Error = Infallible> {
     #[allow(clippy::missing_errors_doc, missing_docs)]
-    fn call(self, input: Input, dummy: &mut Output, emitter: &mut Emitter)
-    -> Result<Output, Error>;
+    fn call(self, input: Input, dummy: &mut Dummy, emitter: &mut Emitter) -> Result<Output, Error>;
 }
 
 macro_rules! impl_attribute_macro {
     ($dummy:ident, $emitter:ident =>
-        $(<$($Inputs:ident),+>(($($in_id:ident),+):($($in_ty:ty),+)$(, $ident:ident:$ty:ty)*);)+) => {
+        $(<$($Inputs:ident $(:$Bound:ident)?),+>(($($in_id:ident),+):($($in_ty:ty),+)$(, $ident:ident:$ty:ty)*), $Dummy:ident;)+) => {
         $(
-        impl<$($Inputs,)+ Output: AnyTokenStream, F> MacroHandler<($($Inputs,)+), Output, ($($in_ty,)* $($ty,)* Output)> for F
+        // NOTE: This `Clone` is just a marker to make this != Emitter
+        impl<$($Inputs $(:$Bound)?,)+ Output: Clone, F> MacroHandler<($($in_ty,)+), $Dummy, Output, ($($in_ty,)* $($ty,)* Output)> for F
         where
             F: FnOnce($($in_ty,)+ $($ty,)*) -> Output
         {
             #[allow(unused)]
-            fn call(self, ($($in_id,)+): ($($in_ty,)+), $dummy: &mut Output, $emitter: &mut Emitter) -> Result<Output, Infallible> {
+            fn call(self, ($($in_id,)+): ($($in_ty,)+), $dummy: &mut $Dummy, $emitter: &mut Emitter) -> Result<Output, Infallible> {
                 Ok(self($($in_id,)+ $($ident),*))
             }
         }
-        impl<$($Inputs,)+ Output: AnyTokenStream, F, Error> MacroHandler<($($Inputs,)+), Output, ($($in_ty,)* $($ty,)* Result<Output, Error>), Error> for F
+        // NOTE: This `Clone` is just a marker to make this != Emitter
+        impl<$($Inputs $(:$Bound)?,)+ Output: Clone, F, Error> MacroHandler<($($in_ty,)+), $Dummy, Output, ($($in_ty,)* $($ty,)* Result<Output, Error>), Error> for F
         where
             F: FnOnce($($in_ty,)+ $($ty,)*) -> Result<Output, Error>
         {
             #[allow(unused)]
-            fn call(self, ($($in_id,)+): ($($in_ty,)+), $dummy: &mut Output, $emitter: &mut Emitter) -> Result<Output, Error> {
+            fn call(self, ($($in_id,)+): ($($in_ty,)+), $dummy: &mut $Dummy, $emitter: &mut Emitter) -> Result<Output, Error> {
                 self($($in_id,)+ $($ident),*)
             }
         }
@@ -504,14 +788,14 @@ macro_rules! impl_attribute_macro {
 
 impl_attribute_macro! {
     dummy, emitter =>
-    <Input, Item> ((input, item): (Input, Item), dummy: &mut Output);
-    <Input, Item> ((input, item): (Input, Item));
-    <Input, Item> ((input, item): (Input, Item), dummy: &mut Output, emitter: &mut Emitter);
-    <Input, Item> ((input, item): (Input, Item), emitter: &mut Emitter);
-    <Input, Item> ((input, item): (Input, Item), emitter: &mut Emitter, dummy: &mut Output);
-    <Input> ((input): (Input), dummy: &mut Output);
-    <Input> ((input): (Input));
-    <Input> ((input): (Input), dummy: &mut Output, emitter: &mut Emitter);
-    <Input> ((input): (Input), emitter: &mut Emitter);
-    <Input> ((input): (Input), emitter: &mut Emitter, dummy: &mut Output);
+    <Input, Item, Dummy: Clone> ((input, item): (Input, Item), dummy: &mut Dummy), Dummy;
+    <Input, Item> ((input, item): (Input, Item)), TokenStream;
+    <Input, Item, Dummy: Clone> ((input, item): (Input, Item), dummy: &mut Dummy, emitter: &mut Emitter), Dummy;
+    <Input, Item> ((input, item): (Input, Item), emitter: &mut Emitter), TokenStream;
+    <Input, Item, Dummy: Clone> ((input, item): (Input, Item), emitter: &mut Emitter, dummy: &mut Dummy), Dummy;
+    <Input, Dummy: Clone> ((input): (Input), dummy: &mut Dummy), Dummy;
+    <Input> ((input): (Input)), TokenStream;
+    <Input, Dummy: Clone> ((input): (Input), dummy: &mut Dummy, emitter: &mut Emitter), Dummy;
+    <Input> ((input): (Input), emitter: &mut Emitter), TokenStream;
+    <Input, Dummy: Clone> ((input): (Input), emitter: &mut Emitter, dummy: &mut Dummy), Dummy;
 }