+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- support `impl Parse` inputs and `impl ToTokens` outputs.
+- added macro alternatives to the `function()`, `derive()` and `attribute()` functions to support `impl Parse/ToTokens`.
+
 ## [0.8.1] - 2023-09-17
 ### Fixed
 - `ensure!(let...)` had compile error in its expansion.